Refactored invictus api calls to own module, plus some code cleanup for better readability
