+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "fundwatch"
+name = "C10"
 version = "0.1.0"
 authors = ["bergabman"]
 edition = "2018"
