+++ src/commands/mod.rs
@@ -1,2 +1,3 @@
 pub mod owner;
 pub mod fund_commands;
+pub mod invictus_api;
